Providing a version of the saxpy function accepting a
vector of hmc_floats as second argument and a saxpby function, to 
make the SolverShifted class work with Wilson spinorfields. Up to now 
only minimal versions are provided to make the code compile. This 
means the function definitions and declarations are given but the 
declarations only contain "throws". The true functionality yet has to 
be implemented. #9