In order to use the function pseudo_randomize with the
rooted staggered spinorfields the method "import" is needed. In order 
to use this method in the rooted staggered spinorfield class it had 
to be put to the protected area in the staggered spinorfield class. 
Furthermore the function pseudo_randomize had to be made a friend of 
the rooted staggered spinorfield class.